main/devicepkg-dev: add test for uefi.config.j2 (MR 4689)
We can't make tests for mainline.config.j2 because it is going to be
modified a lot in the future and maintaining testdata would be crazy.
On the other hand, uefi.config.j2 is pretty small and uses the same
way of checking version and architecture, so testing this file should
be enough as long as we do the same approach in other config.

[ci:skip-build] already built successfully in CI