temp/gtk4.0: new aport
This is a temporary workaround specifially aimed at pine64-pinephone
and similar devices which can't support "recent" graphics standards
by reintroducing the old GL renderer on GTK 4.18.
This happens either as a hardware limitiaion (with the PinePhone) or
a driver limitation (mesa on the Librem 5 isn't capable of that yet).
As we're close to the release I think it's fine to provide this as a
workaround for v25.06 and potentially later.

Convincing upstream to support graphics standards from 2007 is
unlikely, the benefits and maintainability outweigh the complexitiy
of having that upstream. [2]

Thank you @kkofler for making the backport!

[1]: https://linmob.net/gtk-418-the-pinephone-and-megapixels/
[2]: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6548#note_482684