main/devicepkg-dev: kernel_prepare: detect missing deps during validation
This is (hopefully) makes config validation more robust by considering
all fragments together and detecting missing dependencies when an
expected config is not found in the final generated config (which should
hopeqully help reduce time debugging why options are mysteriously not
present)

There are some possible downsides to this change... the logic is more
complex, and probably slower. But in the overall scope of "generating
full kconfig and building a kernel", it's still a tiny drop in the
bucket.