main/postmarketos-ui-console: include zsh
Given that the `console` UI is typically used for an interactive
console, pre-install a more user friendly shell and its completions.

I also propose that we change the default shell for the default user and
ship a minimal configuration to provide a slightly better out-of-the-box
experience with this UI. I'd like to hear some feedback before moving
forward with this.