main/postmarketos-initramfs: add support for immutable booting
This adds support for booting immutable pmOS images, triggered by
the presence of "usrhash=" on the kernel cmdline. This is a safe flag
for "is it immutable?" because it is used by dm-verity for verified
partitions (in this case, the usr partition), so this param is not going
to be present on mutable installs).

Immutable-specific logic, for the most part, is provided by a
separate functions file that is installed by a separate package
(postmarketos-duranium). This functions file is sourced if it exists and
functions to support immutable boot are conditionally.

Some changes were made to subpartition support, since for immutable boot
we need to be able to search PARTUUIDs of subpartitions:
 - find_partition can now locate partitions by PARTUUID in nested
   partition tables using sfdisk (immutable-only, as sfdisk not available
   in mutable initramfs)
 - mount_subpartitions checks for both root and boot partitions to
   support first boot scenarios where root doesn't exist yet