WIP: systemd-services: add tmpfiles configuration for all packages in a plasma-desktop install
As a first step towards
https://gitlab.postmarketos.org/postmarketOS/pmcr/-/blob/main/0002-hermetic-usr.md,
this packages tmpfiles.d configuration for all packages installed by
default in a Qemu amd64 plasma-desktop installation.

The end goal of this commit is to make it possible to do a `rm -r /etc`
after having copied the entirety of it to /usr/share/factory and have
it be repopulated on the next boot automatically by systemd-tmpfiles.