ci: verify_checksums: actually fail check when checksums fail
Due to a quirk in pmb[1], checksum --verify doesn't actually fail if
checksums mismatch when checking local pmaports (IIRC to allow easier
local development), but this causes CI to wrongly pass when checksums fail:

    (native) % chown -R pmos:pmos /home/pmos/build
    (native) verify checksums for device-nothing-tetris
    (native) % cd /home/pmos/build; busybox su pmos -c HOME=/home/pmos abuild fetch verify ;
    >>> device-nothing-tetris: Checking sha512sums...
    deviceinfo: FAILED
    sha512sum: WARNING: 1 of 1 computed checksums did NOT match
    NOTE: chroot is still active (use 'pmbootstrap shutdown' as necessary)
    DONE!
    Running after_script 00:00
    Running after script...
    $ .ci/lib/move_logs.sh $CI_PROJECT_DIR
    WARNING: after_script failed, but job will continue unaffected: exit code 1
    Cleaning up project directory and file based variables 00:01
    Job succeeded

This is bad, because it can cause bpo to get stuck until someone fixes
the checksum.

This patch is basically a hack that looks for "FAILED" in the pmb
output, and throws an exception if it's found.