kconfigcheck: switch many options to modules
Based on running kconfig check against a few kernel configs in pmaports
(which presumably already work/boot), and fixing the INFO messages when
it complained about an option should be builtin.

I also removed some options that were in the default category that were
redefined/duplicated elsewhere, in order to avoid confusion later.