main/postmarketos-bootsplash: kill pbsplash with full privileges
This fixes an issue where pbsplash was not killed when the override is
attached to a service that runs as a non-root user (e.g.
tinydm.service). The tiny `+` tells systemd to run this command with
full privileges, the other stuff in the service's ExecStart and so on
are unaffected by this change.

This also adds the kill pbsplash override to getty@.service, fixes #4018