main/postmarketos-base-ui-plasma: depend on polkit systemd for systemd installations
Since 23e7b54acd488afbf30c666e2b17b8fa2f37970d in aports
polkit-kde-agent-1 doesn't pull in polkit-elogind anymore which seems to
also have caused the systemd installations in postmarketOS to not pull
in polkit-systemd anymore. Since that's a package that should be pulled
in from a meta package anyway (so they can differ per init system
properly) we should pull it in now as otherwise the keyboard input does
not work on SDDM (and there are probably more problems without it as
well).