device-fairphone-fp6: fix useless append_dtb for header v2
This was actually used with boot-deploy < 0.20.0 because
deviceinfo_header_version="2" was ignored and therefore v0 was used,
which requires appended dtb. Now that boot-deploy is actually passing
this parameter, we do not have to append the dtb to the kernel image
anymore, because the dtb is passed in a separate field.