main/postmarketos-mkinitfs-hook-ci: keep dmesg enabled
So we can read the complete output in the logs. This was initially
done to avoid the output from being interrupted, and therfore easier
to read. However, it avoids kernel logs to be collected, which is an
important part of the tests, and necessary for debugging. So let dmesg
be as default.

Signed-off-by: Pablo Correa Gómez <pabloyoyoista@postmarketos.org>