device-pine64-pinenote: rename sway,sway-systemd subpackage to sway-hrdl, systemd
In an effort to have the option to start with a vanilla sway experience,
the sway subpackage is renamed to sway-hrdl to allow for a vanilla sway
config by default, but give the option for hrdl's opiniated
configuration if desired. Non-opiniated eink optimizations will be added to
the now empty sway subpackage as they are developped. A post-install
script was added to move back to vanilla sway.

In favor of using the more automated systemd subpkg. `tinydm-ebc.service` is
removed in favor of an override file for `tinydm.service` packaged in a
device-wide systemd subpackage. A post-install script was added to
disable `tinydm-ebc.service` and enable `tinydm.service` is tinydm is
in-use at time of upgrade.

Enabling systemd units is now handled by postmarketos-base-systemd's
90-base.preset.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7515>
(cherry picked from commit d7229de9d40a381c6992d00d1bdeda5e40fa8332)