main/postmarketos-initramfs: vibrate on boot
Vibrate on multiple stages:
* twice when forced to enter debug-shell
* 3 times on error
* once at the end of the initramfs to indicate that we're booting
  to the OS
This uses the beebzzr tool <https://scm.dersco.re/utils/beebzzr.git>.

It is meant to be useful to know if your phone is booting or having
issues, which is useful when legally or fully blind, or when the display
isn't working yet.

On devices without force-feedback support, this is a no-op.

Signed-off-by: Ferass El Hafidi <funderscore@postmarketos.org>