main/postmarketos-initramfs: vibrate on boot
Vibrate on multiple stages:
* twice when forced to enter debug-shell
* 3 times on error
* once at the end of the initramfs to indicate that we're booting
  to the OS
This uses the beebzzr tool <https://scm.dersco.re/utils/beebzzr.git>.

It is meant to be useful to know in which state the phone is in for
people unable to look at the screen (blind), but may also be useful
in bringup when there's no display yet.

On devices without vibration support, this is a no-op.

Signed-off-by: Ferass El Hafidi <funderscore@postmarketos.org>