githooks/pre-commit: remove file size check
This still comes from the aports hook, where we want to avoid too large patches,
but in pmaports we have big kernel config files that might exceed the
configured 256 kiB limit.

Currently the following files exceed that limit and used git commit
--no-verify to overcome the error (or didn't configure git hooks):

	260K	./device/community/linux-postmarketos-qcom-laptop/config-postmarketos-qcom-laptop.aarch64
	268K	./main/postmarketos-bootsplash/OpenSans-Regular.svg
	336K	./device/downstream/linux-amazon-checkers/0003-Add-missing-kconfig-files.patch
	376K	./device/downstream/linux-amazon-austin/silence-power-logspam.patch

Alternatively we can increase the file limit, but i'm not sure what that
solves.

In future for Alpine, I want to move the check to abuild's internal
linting, which makes it visible for all APKBUILDs, consistent with
other of abuild's linters and also shows up on CI.