device-purism-librem5: disable pipewire
Calls are broken for the Librem 5 with pipewire. Switch to pulseaudio
again until this is resolved by depending on
postmarketosbase-ui-audio-pulseaudio and pulseaudio<9990. The later is
needed as otherwise we get a conflict with pipewire-pulse:

  ERROR: unable to select packages:
    pipewire-pulse-9991.4.9-r1:
      breaks: postmarketos-base-ui-audio-pulseaudio-42-r0[!pipewire-pulse]
      satisfies: device-purism-librem5-9-r6[pulseaudio>17]
                 gnome-settings-daemon-99948.1-r3[pulseaudio]
                 postmarketos-base-ui-audio-pulseaudio-42-r0[pulseaudio]

That pulseaudio<9990 is needed is not great, but doing it like this
inside the device package is the best way to ensure this doesn't affect
any other devices for now, with the intention to have this change
backported to v25.12 before we announce the release. I've created issue
4187 to discuss how to improve the pulseaudio/pipewire packaging in
general so this isn't necessary in the future.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7619>
(cherry picked from commit 613c20c7e76dd2785790411841d2307bb2ae9359)