device-pine64-pinenote: fix chroot install
`device-pine64-pinenote-systemd.post-install` checks if `tinydm-ebc` was
active via `is-active` systemctl command. This does not work in chroot
environment, and exits with exit code 0, thus acting as if `tinydm-ebc`
is active. Since this check is used for v7-r0 to v8-r0 upgrade, it can
be assumed that if output isn't `active`, then the check should not
pass. This commit reworks this by looking for that output rather than a
zero exit code.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7645>